Created check for updating an existing ro-crate and accompanying test…

+++ rocrate/rocrate.py
@@ -465,7 +465,8 @@ class ROCrate():
                 rel = source.relative_to(top)
                 if not self.dereference(str(rel)):
                     dest = base_path / rel
-                    shutil.copyfile(source, dest)
+                    if not dest.exists() or not dest.samefile(source):
+                        shutil.copyfile(source, dest)
 
     def write(self, base_path):
         base_path = Path(base_path)
